Package Authors fix and code clone reduction
