+++ ClippyVSPackage/source.extension.vsixmanifest
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
     <Metadata>
-        <Identity Id="6aae52ac-8840-421f-bb35-60b52ab431e2" Version="0.8" Language="en-US" Publisher="Eros Fratini / Denis Abt" />
+        <Identity Id="6aae52ac-8840-421f-bb35-60b52ab431e2" Version="0.8" Language="en-US" Publisher="Eros Fratini and Denis Abt" />
         <DisplayName>ClippyVS</DisplayName>
         <Description xml:space="preserve">Clippy is back!
 Thanks To Eros Fratini for the initial work and heavy lifting - I've migrated this to be compatible with Visual Studio 2019 &amp; 2022 editions.
