+++ SharedProject1/AssistImpl/Merlin.cs
@@ -71,28 +71,10 @@ MerlinAnimations.Idle1_2};
         /// </summary>
         public Merlin(Panel canvas)
         {
-            var spResUri = SpriteResourceUri;
-#if Dev19
-            spResUri = SpriteResourceUri.Replace("ClippyVs2022", "ClippyVSPackage");
-#endif
-#if Dev22
-#endif
-            this.Sprite = new BitmapImage(new Uri(spResUri, UriKind.RelativeOrAbsolute));
-
-            ClippedImage = new Image
-            {
-                Source = Sprite,
-                Stretch = Stretch.None
-            };
-
-            if (canvas == null) return;
-
-            canvas.Children.Clear();
-            canvas.Children.Add(ClippedImage);
+            InitAssistant(canvas, SpriteResourceUri);
 
             if (_animations == null)
                 RegisterAnimations();
-
 
             //XX Requires testing..
             AllAnimations = new List<MerlinAnimations>();
@@ -106,74 +88,7 @@ MerlinAnimations.Idle1_2};
         /// </summary>
         private void RegisterAnimations()
         {
-            var spResUri = AnimationsResourceUri;
-
-#if Dev19
-            spResUri = spResUri.Replace("ClippyVs2022", "ClippyVSPackage");
-#endif
-            var uri = new Uri(spResUri, UriKind.RelativeOrAbsolute);
-
-            var info = Application.GetResourceStream(uri);
-
-            if (info == null) 
-                return;
-
-            // Can go to Constructor/Init
-            var storedAnimations = 
-                Newtonsoft.Json.JsonConvert.DeserializeObject<List<ClippySingleAnimation>>(StreamToString(info.Stream));
-
-            _animations = new Dictionary<string, Tuple<DoubleAnimationUsingKeyFrames, DoubleAnimationUsingKeyFrames>>();
-
-            if (storedAnimations == null) return;
-
-            foreach (var animation in storedAnimations)
-            {
-                var xDoubleAnimation = new DoubleAnimationUsingKeyFrames
-                {
-                    FillBehavior = FillBehavior.HoldEnd
-                };
-
-                var yDoubleAnimation = new DoubleAnimationUsingKeyFrames
-                {
-                    FillBehavior = FillBehavior.HoldEnd
-                };
-
-                double timeOffset = 0;
-
-                foreach (var frame in animation.Frames)
-                {
-                    if (frame.ImagesOffsets != null)
-                    {
-                        var lastCol = frame.ImagesOffsets.Column;
-                        var lastRow = frame.ImagesOffsets.Row;
-
-                        // X
-                        var xKeyFrame = new DiscreteDoubleKeyFrame(ClipWidth * -lastCol, KeyTime.FromTimeSpan(TimeSpan.FromSeconds(timeOffset)));
-
-                        // Y
-                        var yKeyFrame = new DiscreteDoubleKeyFrame(ClipHeight * -lastRow, KeyTime.FromTimeSpan(TimeSpan.FromSeconds(timeOffset)));
-
-                        timeOffset += ((double)frame.Duration / 1000);
-                        xDoubleAnimation.KeyFrames.Add(xKeyFrame);
-                        yDoubleAnimation.KeyFrames.Add(yKeyFrame);
-                    }
-                    else
-                    {
-                        Debug.WriteLine("ImageOffsets was null");
-                    }
-                }
-
-                _animations.Add(animation.Name, new Tuple<DoubleAnimationUsingKeyFrames, DoubleAnimationUsingKeyFrames>(xDoubleAnimation, yDoubleAnimation));
-                Debug.WriteLine("Added Merlin Anim {0}", animation.Name);
-
-                xDoubleAnimation.Changed += XDoubleAnimation_Changed;
-                xDoubleAnimation.Completed += XDoubleAnimation_Completed;
-            }
-        }
-
-        private void XDoubleAnimation_Changed(object sender, EventArgs e)
-        {
-            //Debug.WriteLine("Merlin: Animation changing");
+            RegisterAnimationsImpl(AnimationsResourceUri, ref _animations, XDoubleAnimation_Completed, ClipWidth, ClipHeight);
         }
 
         /// <summary>
